+++ src/main/java/net/rptools/maptool/client/ui/MacroButtonDialog.java
@@ -22,6 +22,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.HashSet;
 import javax.swing.Action;
@@ -47,6 +49,7 @@ import net.rptools.lib.swing.SwingUtil;
 import net.rptools.lib.swing.preference.WindowPreferences;
 import net.rptools.maptool.client.AppActions;
 import net.rptools.maptool.client.AppConstants;
+import net.rptools.maptool.client.AppPreferences;
 import net.rptools.maptool.client.AppUtil;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.MapToolUtil;
@@ -68,6 +71,7 @@ import org.fife.ui.autocomplete.CompletionProvider;
 import org.fife.ui.rsyntaxtextarea.AbstractTokenMakerFactory;
 import org.fife.ui.rsyntaxtextarea.ErrorStrip;
 import org.fife.ui.rsyntaxtextarea.RSyntaxTextArea;
+import org.fife.ui.rsyntaxtextarea.SyntaxConstants;
 import org.fife.ui.rsyntaxtextarea.Theme;
 import org.fife.ui.rsyntaxtextarea.TokenMakerFactory;
 import org.fife.ui.rsyntaxtextarea.folding.CurlyFoldParser;
@@ -352,6 +356,7 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
             .setSelected(properties.getCompareApplyToSelectedTokens());
         getAllowPlayerEditsCheckBox().setSelected(properties.getAllowPlayerEdits());
         getToolTipTextField().setText(properties.getToolTip());
+        getDisplayHotkeyCheckBox().setSelected(properties.getDisplayHotKey());
 
         if (isCommonMacro) {
           getColorComboBox().setEnabled(false);
@@ -389,6 +394,14 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
         (AbstractTokenMakerFactory) TokenMakerFactory.getDefaultInstance();
     atmf.putMapping(
         "text/MapToolScript", "net.rptools.maptool.client.ui.syntax.MapToolScriptSyntax");
+
+    // Expanding use of tooltip - already accepts HTML so lets show it
+    getToolTipTextField().setSyntaxEditingStyle(SyntaxConstants.SYNTAX_STYLE_HTML);
+    getToolTipTextField().setLineWrap(true);
+    getToolTipTextField().setWrapStyleWord(true);
+    getToolTipTextField().setTabSize(2);
+
+    // Macro Editor setup
     macroEditorRSyntaxTextArea.setSyntaxEditingStyle("text/MapToolScript");
 
     macroEditorRSyntaxTextArea.setEditable(true);
@@ -406,27 +419,17 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
     ac.setAutoActivationEnabled(true);
     ac.setAutoActivationDelay(500);
     ac.setShowDescWindow(true);
+    ac.setAutoCompleteSingleChoices(false);
     ac.install(macroEditorRSyntaxTextArea);
 
     // Set the color style via Theme
     try {
-      // Theme theme =
-      // Theme.load(getClass().getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/default.xml"));
-      // Theme theme =
-      // Theme.load(getClass().getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/dark.xml"));
-      // Theme theme =
-      // Theme.load(getClass().getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/eclipse.xml"));
-      // Theme theme =
-      // Theme.load(getClass().getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/idea.xml"));
-      // Theme theme =
-      // Theme.load(getClass().getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/vs.xml"));
-      Theme theme =
-          Theme.load(
-              getClass()
-                  .getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/nerps.xml"));
-      // Theme theme =
-      // Theme.load(getClass().getResourceAsStream("/net/rptools/maptool/client/ui/syntax/themes/nerps-dark.xml"));
+      File themeFile =
+          new File(AppConstants.THEMES_DIR, AppPreferences.getDefaultMacroEditorTheme() + ".xml");
+      Theme theme = Theme.load(new FileInputStream(themeFile));
       theme.apply(macroEditorRSyntaxTextArea);
+      theme.apply(getToolTipTextField());
+
       macroEditorRSyntaxTextArea.revalidate();
     } catch (IOException e) {
       e.printStackTrace();
@@ -453,13 +456,18 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
     csp = new CollapsibleSectionPanel();
     ((GridView) panel.getComponentByName("macroEditorPanel")).add(csp);
 
-    ErrorStrip errorStrip = new ErrorStrip(macroEditorRSyntaxTextArea);
-    csp.add(errorStrip, BorderLayout.LINE_END);
+    csp.add(new ErrorStrip(macroEditorRSyntaxTextArea), BorderLayout.LINE_END);
 
     RTextScrollPane macroEditorRTextScrollPane = new RTextScrollPane(macroEditorRSyntaxTextArea);
     macroEditorRTextScrollPane.setLineNumbersEnabled(true);
     // replaceComponent("macroEditorPanel", "macroEditorRTextScrollPane",
     // macroEditorRTextScrollPane);
+
+    csp.add(new ErrorStrip(getToolTipTextField()), BorderLayout.LINE_END);
+
+    //    RTextScrollPane macroEditorRTextScrollPane = new
+    // RTextScrollPane(macroEditorRSyntaxTextArea);
+    //    macroEditorRTextScrollPane.setLineNumbersEnabled(true);
 
     csp.add(macroEditorRTextScrollPane);
   }
@@ -497,7 +505,7 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
     JMenuBar mb = MapTool.getFrame().getJMenuBar();
     for (int i = 0; i < mb.getMenuCount(); i++) {
       JMenu menu = mb.getMenu(i);
-      if (menu.getText().equalsIgnoreCase("Edit")) {
+      if (menu.getText().equalsIgnoreCase(I18N.getText("menu.edit"))) {
         // This is the menu we want to add onto...
         addMenuItems(menu);
         return;
@@ -544,7 +552,7 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
     JMenuBar mb = MapTool.getFrame().getJMenuBar();
     for (int i = 0; i < mb.getMenuCount(); i++) {
       JMenu menu = mb.getMenu(i);
-      if (menu.getText().equalsIgnoreCase("Edit")) {
+      if (menu.getText().equalsIgnoreCase(I18N.getText("menu.edit"))) {
         // This is the menu we want to cleanup...
         removeMenuItems(menu, menu.getItemCount());
         return;
@@ -686,6 +694,7 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
         getCompareApplyToSelectedTokensCheckBox().isSelected());
     properties.setAllowPlayerEdits(getAllowPlayerEditsCheckBox().isSelected());
     properties.setToolTip(getToolTipTextField().getText());
+    properties.setDisplayHotKey(getDisplayHotkeyCheckBox().isSelected());
 
     properties.save();
 
@@ -850,8 +859,12 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
     return panel.getCheckBox("allowPlayerEditsCheckBox");
   }
 
-  private JTextField getToolTipTextField() {
-    return panel.getTextField("toolTip");
+  private JCheckBox getDisplayHotkeyCheckBox() {
+    return panel.getCheckBox("displayHotKeyCheckBox");
+  }
+
+  private RSyntaxTextArea getToolTipTextField() {
+    return (RSyntaxTextArea) panel.getComponentByName("toolTip");
   }
 
   // Begin comparison customization
@@ -937,6 +950,9 @@ public class MacroButtonDialog extends JDialog implements SearchListener {
     getAllowPlayerEditsCheckBox().setText(I18N.getText("component.label.macro.allowPlayerEdits"));
     getAllowPlayerEditsCheckBox()
         .setToolTipText(I18N.getText("component.tooltip.macro.allowPlayerEdits"));
+    getDisplayHotkeyCheckBox().setText(I18N.getText("component.label.macro.displayHotKey"));
+    getDisplayHotkeyCheckBox()
+        .setToolTipText(I18N.getText("component.tooltip.macro.displayHotKey"));
     ((TitledBorder) ((GridView) panel.getComponentByName("macroComparisonGridView")).getBorder())
         .setTitle(I18N.getText("component.label.macro.macroCommonality"));
     getCompareIncludeLabelCheckBox()