+++ src/main/java/net/rptools/maptool/model/HexGrid.java
@@ -228,7 +228,10 @@ public abstract class HexGrid extends Grid {
   protected Area createCellShape(int size) {
     // don't use size. it has already been used to set the minorRadius
     // and will only introduce a rounding error.
-    return new Area(createShape(minorRadius, edgeProjection, edgeLength));
+    Area a = new Area(createShape(minorRadius, edgeProjection, edgeLength));
+    //    System.out.println("HexGrid.createCellShape(): " + a.getBounds().width + ":" +
+    // a.getBounds().height);
+    return a;
   }
 
   @Override
@@ -289,8 +292,6 @@ public abstract class HexGrid extends Grid {
 
   @Override
   public void setSize(int size) {
-    super.setSize(size);
-
     if (hexRatio == 0) {
       hexRatio = REGULAR_HEX_RATIO;
     }
@@ -308,6 +309,9 @@ public abstract class HexGrid extends Grid {
 
     // Cell offset gives the offset to apply to the cell zone coords to draw images/tokens
     cellOffset = setCellOffset();
+    // The call to the super.setSize() must be last as it calls createCellShape()
+    // which needs the values set above.
+    super.setSize(size);
   }
 
   protected void createShape(double scale) {