+++ src/main/java/net/rptools/maptool/model/Token.java
@@ -467,7 +467,11 @@ public class Token extends BaseModel implements Cloneable {
   }
 
   public String getGMNotes() {
-    return gmNotes;
+    if (MapTool.getPlayer().isGM()) {
+      return gmNotes;
+    } else {
+      return "";
+    }
   }
 
   public void setGMNotes(String notes) {
@@ -475,7 +479,11 @@ public class Token extends BaseModel implements Cloneable {
   }
 
   public String getGMName() {
-    return gmName;
+    if (MapTool.getPlayer().isGM()) {
+      return gmName;
+    } else {
+      return "";
+    }
   }
 
   public void setGMName(String name) {
@@ -626,17 +634,31 @@ public class Token extends BaseModel implements Cloneable {
     this.facing = facing;
   }
 
+  /**
+   * Facing is in the map space where 0 degrees is along the X axis to the right and proceeding CCW
+   * for positive angles.
+   *
+   * <p>Round/Square tokens that have no facing set, return null. Top Down tokens default to -90.
+   *
+   * @return null or angle in degrees
+   */
   public Integer getFacing() {
     return facing;
   }
 
+  /**
+   * This returns the rotation of the facing of the token from the default facing of down or -90.
+   * Positive for CW and negative for CCW.
+   *
+   * @return angle in degrees
+   */
   public Integer getFacingInDegrees() {
-    if (facing == null) return -1;
+    if (facing == null) return 0;
     else return -(facing + 90);
   }
 
   public Integer getFacingInRealDegrees() {
-    if (facing == null) return -1;
+    if (facing == null) return 270;
 
     if (facing >= 0) return facing;
     else return facing + 360;
@@ -1206,12 +1228,6 @@ public class Token extends BaseModel implements Cloneable {
     TokenFootprint footprint = getFootprint(grid);
     Rectangle footprintBounds =
         footprint.getBounds(grid, grid.convert(new ZonePoint(getX(), getY())));
-    // if (getShape() == TokenShape.FIGURE) {
-    // double th = this.height * Double.valueOf(footprintBounds.width) / this.width;
-    // double ho = footprintBounds.height - th;
-    // footprintBounds = new Rectangle(footprintBounds.x, footprintBounds.y + (int)ho,
-    // footprintBounds.width, (int)th);
-    // }
 
     double w = footprintBounds.width;
     double h = footprintBounds.height;
@@ -1220,6 +1236,10 @@ public class Token extends BaseModel implements Cloneable {
     if (!isSnapToScale()) {
       w = getWidth() * getScaleX();
       h = getHeight() * getScaleY();
+      if (grid.isIsometric() && getShape() == Token.TokenShape.FIGURE) {
+        // Native size figure tokens need to follow iso rules
+        h = (w / 2);
+      }
     } else {
       w = footprintBounds.width * footprint.getScale() * sizeScale;
       h = footprintBounds.height * footprint.getScale() * sizeScale;